requirements.txt
pandas
numpy
python-dotenv
requests
plotly
dash
dash_bootstrap_components
supabase
